training MINIROCKET classifier on financial data
ohlcv timeseries' are augmented with semi-randomized ses of technical
indicators and are then used to fit a MINIROCKET-classifier.
As of this commit, the models are only actually trained on 17% of the data for a
given symbol, and are then evaluated on the remaining 83%, which makes
the fact that many of the models evaluate at >90% accuracy pretty
goddamn impressive!

Next, backtest to measure how profitable such accuracy can be made, and
clean up this horrendous f*cking codebase :3